+++ ush/wave_tar.sh
@@ -1,4 +1,5 @@
-#!/bin/bash
+#! /usr/bin/env bash
+
 ###############################################################################
 #                                                                             #
 # This script tars the sectral or bulletin files into a single file and       #
@@ -23,15 +24,13 @@
 #
 # --------------------------------------------------------------------------- #
 # 0.  Preparations
+
+PREAMBLE_SCRIPT="${PREAMBLE_SCRIPT:-$HOMEgfs/ush/preamble.sh}"
+if [ -f "${PREAMBLE_SCRIPT}" ]; then
+  source $PREAMBLE_SCRIPT
+fi
+
 # 0.a Basic modes of operation
-
-  # set execution trace prompt.  ${0##*/} adds the script's basename
-  PS4=" \${SECONDS} ${0##*/} L\${LINENO} + "
-  set -x
-
-  # Use LOUD variable to turn on/off trace.  Defaults to YES (on).
-  export LOUD=${LOUD:-YES}; [[ $LOUD = yes ]] && export LOUD=YES
-  [[ "$LOUD" != YES ]] && set -x
 
   cd $DATA
   echo "Making TAR FILE"
@@ -46,7 +45,7 @@
   echo "   ID              : $1"
   echo "   Type            : $2"
   echo "   Number of files : $3"
-  [[ "$LOUD" = YES ]] && set -x
+  ${TRACE_ON:-set -x}
 
 
 # 0.b Check if type set
@@ -59,7 +58,7 @@
     echo '*** VARIABLES IN ww3_tar.sh NOT SET ***'
     echo '********************************************'
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
+    ${TRACE_ON:-set -x}
     exit 1
   else
     ID=$1
@@ -89,7 +88,7 @@
     echo '*** EXPORTED VARIABLES IN ww3_tar.sh NOT SET ***'
     echo '*****************************************************'
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
+    ${TRACE_ON:-set -x}
     exit 2
   fi
 
@@ -101,7 +100,7 @@
   set +x
   echo ' '
   echo '   Making tar file ...'
-  set -x
+  ${TRACE_ON:-set -x}
 
   count=0
   countMAX=5
@@ -125,7 +124,7 @@
         echo '*** FATAL ERROR : TAR CREATION FAILED *** '
         echo '***************************************** '
         echo ' '
-        [[ "$LOUD" = YES ]] && set -x
+        ${TRACE_ON:-set -x}
         exit 3
       fi
       
@@ -136,7 +135,7 @@
     else
       set +x
       echo ' All files not found for tar. Sleeping 10 seconds and trying again ..'
-      [[ "$LOUD" = YES ]] && set -x
+      ${TRACE_ON:-set -x}
       sleep 10
       count=$(expr $count + 1)
     fi
@@ -151,7 +150,7 @@
     echo '*** FATAL ERROR : TAR CREATION FAILED *** '
     echo '***************************************** '
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
+    ${TRACE_ON:-set -x}
     exit 3
   fi
 
@@ -171,7 +170,7 @@
         echo '*** FATAL ERROR : SPECTRAL TAR COMPRESSION FAILED *** '
         echo '***************************************************** '
         echo ' '
-        [[ "$LOUD" = YES ]] && set -x
+        ${TRACE_ON:-set -x}
         exit 4
       fi
     fi
@@ -185,7 +184,7 @@
   set +x
   echo ' '
   echo "   Moving tar file ${file_name} to $COMOUT ..."
-  [[ "$LOUD" = YES ]] && set -x
+  ${TRACE_ON:-set -x}
 
   cp ${file_name} $COMOUT/station/.
 
@@ -199,7 +198,7 @@
     echo '*** FATAL ERROR : TAR COPY FAILED *** '
     echo '************************************* '
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
+    ${TRACE_ON:-set -x}
     exit 4
   fi
 
@@ -209,21 +208,19 @@
     echo ' '
     echo "   Alerting TAR file as $COMOUT/station/${file_name}"
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
+    ${TRACE_ON:-set -x}
     $DBNROOT/bin/dbn_alert MODEL ${alertName}_WAVE_TAR $job $COMOUT/station/${file_name}
   fi
 
 # --------------------------------------------------------------------------- #
 # 4.  Final clean up
 
-  cd $DATA
+cd $DATA
 
-  set +x; [[ "$LOUD" = YES ]] && set -v
+if [[ ${KEEPDATA:-NO} == "NO" ]]; then
+  set -v
   rm -rf  ${STA_DIR}/${type}
   set +v
-
-  echo ' '
-  echo 'End of ww3_tar.sh at'
-  date
+fi
 
 # End of ww3_tar.sh ----------------------------------------------------- #
